requirements.txt
beautifulsoup4==4.4.0
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-environ==0.4.4
django-filter==2.0.0.dev1
django-rest-swagger==0.3.10
djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
Markdown==2.6.9
psycopg2==2.7.3.2
pytz==2017.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
